+++ src/QueryBuilder/Condition/NotIn.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Db\QueryBuilder\Condition;
 
 /**
- * Condition that represents `IN` operator.
+ * Condition that represents `NOT IN` operator.
  */
-final class In extends AbstractIn
+final class NotIn extends AbstractIn
 {
 }